Add merge suggestions ux fixes
